webutil common moved to webmain
