+++ sample/src/jsMain/kotlin/Video.kt
@@ -24,17 +24,16 @@ import com.hamama.kwhi.HtmlView
 import external.videojs
 import kotlinx.browser.document
 import kotlinx.browser.window
-import modals.VideoInitOptionsModal
 import modals.VideoPlayer
-import modals.VideoSource
-import util.isVideoJsFuncAvailable
 import util.loadCss
 import util.loadJs
-import util.newJsObject
 import utils.VIDEO_JS_CSS_ID
 import utils.VIDEO_JS_CSS_URL
 import utils.VIDEO_JS_ID
 import utils.VIDEO_JS_URL
+import utils.createVideoOptionsObject
+import utils.createVideoSource
+import utils.isVideoJsFuncAvailable
 
 @Composable
 fun App() {
@@ -179,26 +178,4 @@ fun listenErrorEvents(player: VideoPlayer) {
         println("'Internet connection present!'")
         player.play()
     })
-}
-
-
-fun createVideoOptionsObject(videoSrcUrl: String, videoType: String): VideoInitOptionsModal {
-    val source = createVideoSource(videoSrcUrl, videoType)
-
-    val videoOption = newJsObject<VideoInitOptionsModal>().apply {
-        this.controls = true
-        this.autoplay = false
-        this.preload = "auto"
-        this.sources = arrayOf(source)
-    }
-    return videoOption
-}
-
-
-fun createVideoSource(videoSrcUrl: String, videoType: String): VideoSource {
-    val source = newJsObject<VideoSource>().apply {
-        this.src = videoSrcUrl
-        this.type = videoType
-    }
-    return source
 }
